ChangePitchBase in lib-builtin-effects
